# flag-backend-master/src/main/resources/META-INF/openapi.yaml
openapi: 3.0.0
info:
  title: DCTx FLAG API
  description: >-
    API for FLAG Web Services
  version: "1.0.0"
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
security:
  - BearerAuth: []
paths:
...
  /static/currencies:
    get:
      tags:
        - static
      summary: >-
        Retrieve all currencies
      operationId: getCurrencies
      responses:
        200:
          description: >-
            All currencies
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /static/donationtypes:
    get:
      tags:
        - static
      summary: >-
        Retrieve all types of donations
      operationId: getDonationTypes
      responses:
        200:
          description: >-
            All donation types
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /static/donorsectors:
    get:
      tags:
        - static
      summary: >-
        Retrieve all types of donor sectors
      operationId: getDonorSectors
      responses:
        200:
          description: >-
            All donor sectors
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /static/units:
    get:
      tags:
        - static
      summary: >-
        Retrieve all units
      operationId: getUnits
      responses:
        200:
          description: >-
            All units
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /static/servicetypes:
    get:
      tags:
        - static
      summary: >-
        Retrieve all service types
      operationId: getServiceType
      responses:
        200:
          description: >-
            All service types
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /static/beneficiarysectors:
    get:
      tags:
        - static
      summary: >-
        Retrieve all beneficiary sectors
      operationId: getBenefeciarySectors
      responses:
        200:
          description: >-
            All beneficiary sectors
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /static/requesttypes:
    get:
      tags:
        - static
      summary: >-
        Retrieve all request types
      operationId: getRequestTypes
      responses:
        200:
          description: >-
            All request types
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /static/items:
    get:
      tags:
        - static
      summary: >-
        Retrieve all donation items
      operationId: getDonationItems
      responses:
        200:
          description: >-
            All donation items
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /static/regions:
    get:
      tags:
        - static
      summary: >-
        Retrieve all regoins
      operationId: getRegions
      responses:
        200:
          description: >-
            All regions
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /static/{region}/provinces:
    get:
      tags:
        - static
      summary: >-
        Retrieve all provinces by region
      operationId: getProvincesByRegion
      parameters:
        - name: region
          in: path
          required: true
          schema:
            type: string
            description: the region
            example: 'Region VII'
      responses:
        200:
          description: >-
            All provinces in a region
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /static/provinces:
    get:
      tags:
        - static
      summary: >-
        Retrieve all provinces
      operationId: getProvinces
      responses:
        200:
          description: >-
            All provinces
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /static/municipalities:
    get:
      tags:
        - static
      summary: >-
        Retrieve all municipalities
      operationId: getMunicipalities
      responses:
        200:
          description: >-
            All municipalities
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /static/{province}/municipalities:
    get:
      tags:
        - static
      summary: >-
        Retrieve all municipalities in a province
      operationId: getMunicipalitiesByProvince
      parameters:
        - name: province
          in: path
          required: true
          schema:
            type: string
            description: the province
            example: 'Cebu'
      responses:
        200:
          description: >-
            All municipalities in a province
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /static/countries:
    get:
      tags:
        - static
      summary: >-
        Retrieve all countries
      operationId: getCountries
      responses:
        200:
          description: >-
            All regions
        default:
          description: >-
            An error occured while performing the operation. A details JSON error is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        id:
          type: string
          description: >-
            a unique identifier for this particular occurrence of the problem.
        links:
          type: object
          description: >-
            a links object containing the following members;
          properties:
            about:
              type: string
              description: >-
                a link that leads to further details about this particular occurrence of the problem.
          additionalProperties: true
        status:
          type: string
          description: >-
            the HTTP status code applicable to this problem, expressed as a
            string value.
        code:
          type: string
          description: >-
            an application-specific error code, expressed as a string value.
        title:
          type: string
          description: >-
            a short, human-readable summary of the problem that SHOULD NOT
            change from occurrence to occurrence of the problem, except for
            purposes of localization.
        detail:
          type: string
          description: >-
            a human-readable explanation specific to this occurrence of the
            problem. Like title, this field’s value can be localized.
        source:
          type: array
          description: >-
            an object containing references to the source of the error, optionally including any of the following members;
          items:
            type: object
            properties:
              pointer:
                type: string
                description: >-
                  a JSON Pointer [RFC6901] to the associated entity in the
                  request document [e.g. "/data" for a primary data object, or "
                  /data/attributes/title" for a specific attribute].
              parameter:
                type: string
                description: >-
                  a string indicating which URI query parameter caused the error
                  .
        meta:
          type: object
          description: >-
            a meta object containing non-standard meta-information about the error.
          additionalProperties: true
...